+++ src/main/java/com/rm/myadmin/services/ResidenceService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.rm.myadmin.entities.Residence;
 import com.rm.myadmin.repositories.ResidenceRepository;
+import com.rm.myadmin.services.exceptions.DatabaseException;
 import com.rm.myadmin.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -29,6 +32,41 @@ public class ResidenceService {
 	}
 
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (repository.existsById(id)) {
+				repository.deleteById(id);
+			} else {
+				throw new ResourceNotFoundException(id);
+			}
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
+
+	public Residence update(Long id, Residence obj) {
+		Residence entity = repository.getReferenceById(id);
+		updateData(entity, obj);
+		return repository.save(entity);
+	}
+
+	private void updateData(Residence entity, Residence obj) {
+		entity.setPropertyType(obj.getPropertyType());
+		entity.setDescription(obj.getDescription());
+		entity.setAptNumber(obj.getAptNumber());
+		entity.setComplement(obj.getComplement());
+		entity.setNumberBedrooms(obj.getNumberBedrooms());
+		entity.setNumberBathrooms(obj.getNumberBathrooms());
+		entity.setNumberSuites(obj.getNumberSuites());
+		entity.setTotalArea(obj.getTotalArea());
+		entity.setBuiltArea(obj.getBuiltArea());
+		entity.setGarageSpaces(obj.getGarageSpaces());
+		entity.setYearConstruction(obj.getYearConstruction());
+		entity.setOccupancyStatus(obj.getOccupancyStatus());
+		entity.setMarketValue(obj.getMarketValue());
+		entity.setRentalValue(obj.getRentalValue());
+		entity.setDateLastRenovation(obj.getDateLastRenovation());
+	}
+
 }
